feat: move result tools to test_toolbox.py
