+++ tests/test_toolbox.py
@@ -1,5 +1,6 @@
 from functools import partial
 import numpy as np
+import pandas as pd
 
 
 class ParameterTools(object):
@@ -9,8 +10,8 @@ class ParameterTools(object):
             attr
             for attr in dir(item)
             if not callable(getattr(item, attr))
-            and not attr.startswith("_")
-            and not isinstance(getattr(item, attr), partial)
+               and not attr.startswith("_")
+               and not isinstance(getattr(item, attr), partial)
         ]
 
     @staticmethod
@@ -66,3 +67,44 @@ class ParameterTools(object):
                 )
             else:
                 assert getattr(params1, attr) == getattr(params2, attr)
+
+
+class ResultTools(object):
+    @staticmethod
+    def compare_lists_of_floats(list_1: list, list_2: list):
+        """Compares lists of floats"""
+        list_1 = [round(element, 10) for element in list_1]
+        list_2 = [round(element, 10) for element in list_2]
+        assert list_1 == list_2
+
+    def save_spectrum_to_excel(self, array_result, out_excel, result):
+        for idx in np.ndindex(array_result.shape[:-2]):
+            spectrum = array_result[idx]
+            result.spectrum.update({idx: spectrum})
+
+        # if out_excel.is_file():
+        #     out_excel.unlink()
+        bins = np.linspace(0, 10, 11)
+        result.save_spectrum_to_excel(out_excel, bins)
+        df = pd.read_excel(out_excel, index_col=0)
+        columns = df.columns.tolist()
+        assert columns == ["pixel"] + bins.tolist()
+        for idx, key in enumerate(result.spectrum.keys()):
+            spectrum = np.array(df.iloc[idx, 1:])
+            self.compare_lists_of_floats(
+                spectrum.tolist(), np.squeeze(result.spectrum[key]).tolist()
+            )
+
+    def save_curve_to_excel(self, array_result, out_excel, result):
+        for idx in np.ndindex(array_result.shape[:-2]):
+            curve = array_result[idx]
+            result.curve.update({idx: curve})
+
+        b_values = np.linspace(0, 10, 11).tolist()
+        result.save_fit_curve_to_excel(out_excel, b_values)
+        df = pd.read_excel(out_excel, index_col=0)
+        columns = df.columns.tolist()
+        assert columns == ["pixel"] + b_values
+        for idx, key in enumerate(result.curve.keys()):
+            curve = np.array(df.iloc[idx, 1:])
+            self.compare_lists_of_floats(curve.tolist(), np.squeeze(result.curve[key].tolist()))
